[level 0] Title: 구슬을 나누는 경우의 수, Time: 1.23 ms, Memory: 73 MB -BaekjoonHub

+++ 프로그래머스/lv0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.java
@@ -4,7 +4,7 @@ class Solution {
     public BigInteger solution(int balls, int share) {
         
         
-        long bs = balls - share;
+        int bs = balls - share;
         BigInteger b = BigInteger.ONE;
         BigInteger s = BigInteger.ONE;
         BigInteger c = BigInteger.ONE;
